fix issue #14679: correct shape for vec*mat error with test case
In addition, macro except_str is moved from test/replutil.jl to base/test.jl for more usage in tests.
