+++ base/linalg/matmul.jl
@@ -85,7 +85,15 @@ function (*){T,S}(A::AbstractMatrix{T}, x::AbstractVector{S})
     TS = promote_op(MulFun(),arithtype(T),arithtype(S))
     A_mul_B!(similar(x,TS,size(A,1)),A,x)
 end
-(*)(A::AbstractVector, B::AbstractMatrix) = reshape(A,length(A),1)*B
+function (*)(A::AbstractVector, B::AbstractMatrix)
+    mB, nB = size(B)
+    if mB != 1
+        mA = size(A)
+        throw(DimensionMismatch("A has dimensions $mA but B has dimensions ($mB,$nB)"))
+    else
+        reshape(A,length(A),1)*B
+    end
+end
 
 A_mul_B!{T<:BlasFloat}(y::StridedVector{T}, A::StridedVecOrMat{T}, x::StridedVector{T}) = gemv!(y, 'N', A, x)
 for elty in (Float32,Float64)
